Move conversions and roll to their own cogs
