+++ koabot/koakuma.py
@@ -18,12 +18,9 @@ import mysql.connector as mariadb
 import pixivpy3
 import tweepy
 from discord.ext import commands
-from num2words import num2words
 
-import koabot.converter
 import koabot.tasks
 import koabot.utils
-from koabot.patterns import *
 
 bot = commands.Bot(command_prefix='!', description='')
 SOURCE_DIR = os.path.dirname(os.path.realpath(__file__))
@@ -36,103 +33,6 @@ os.makedirs(CONFIG_DIR, exist_ok=True)
 # Make cache dir if it doesn't exist
 CACHE_DIR = appdirs.user_cache_dir(BOT_DIRNAME)
 os.makedirs(CACHE_DIR, exist_ok=True)
-
-
-@bot.command()
-async def roll(ctx, *dice):
-    """Rolls one or many dice"""
-
-    if len(dice) < 1:
-        await ctx.send('Please specify what you want to roll.')
-        return
-
-    dice_matches = re.findall(DICE_PATTERN, ' '.join(dice))
-
-    if not dice_matches:
-        await ctx.send('Sorry, I can\'t do that...')
-        return
-
-    dice_single_or_many = len(dice_matches) > 1 or (dice_matches[0][0] and int(dice_matches[0][0]) > 1)
-    message = '>>> {} rolled the {}.\n'.format(ctx.author.mention, dice_single_or_many and 'dice' or 'die')
-    pip_sum = 0
-
-    for match in dice_matches:
-        quantity = match[0] and int(match[0]) or 1
-        pips = match[1] and int(match[1]) or 1
-        bonus_points = match[2] and int(match[2]) or 0
-
-        message += '{} {}-sided {} for a '.format(num2words(quantity).capitalize(), pips, quantity > 1 and 'dice' or 'die')
-
-        for i in range(0, quantity):
-            die_roll = random.randint(1, pips)
-
-            if i == quantity - 1:
-                if quantity == 1:
-                    message += '{}.'.format(die_roll)
-                else:
-                    message += 'and a {}.'.format(die_roll)
-
-                if bonus_points:
-                    pip_sum += bonus_points
-
-                    if bonus_points > 0:
-                        message += ' +{}'.format(bonus_points)
-                    else:
-                        message += ' {}'.format(bonus_points)
-
-                message += '\n'
-            elif i == quantity - 2:
-                message += '{} '.format(die_roll)
-            else:
-                message += '{}, '.format(die_roll)
-
-            pip_sum += die_roll
-
-    message += 'For a total of **{}.**'.format(pip_sum)
-
-    await ctx.send(message)
-
-
-@bot.command(name='convert', aliases=['conv', 'cv'])
-async def unit_convert(ctx, *, units):
-    """Convert units"""
-
-    unit_matches = []
-    i = 0
-    while i < len(units):
-        ftin_match = SPECIAL_UNIT_PATTERN_TUPLE[1].match(units, i)
-        if ftin_match:
-            unit_matches.append((SPECIAL_UNIT_PATTERN_TUPLE[0], float(ftin_match.group(1)), float(ftin_match.group(2))))
-            # unit_matches.append((unit_name, value in feet, value in inches))
-            i = ftin_match.end()
-            continue
-
-        num_match = NUMBER_PATTERN.match(units, i)
-        if num_match:
-            i = num_match.end()
-            def match(u): return (u[0], u[1].match(units, i))
-            def falsey(x): return not x[1]
-            unit = next(itertools.dropwhile(falsey, map(match, iter(UNIT_PATTERN_TUPLE))), None)
-            if unit:
-                (unit, unit_match) = unit
-                unit_matches.append((unit, float(num_match.group(1))))
-                i = unit_match.end()
-
-        i += 1
-
-    if unit_matches:
-        await koabot.converter.convert_units(ctx, unit_matches)
-
-
-@bot.command(name='exchange', aliases=['currency', 'xc', 'c'])
-async def convert_currency(ctx, amount, currency_type1, _, currency_type2):
-    """Convert currency to others"""
-
-    currency_type1 = currency_type1.upper()
-    currency_type2 = currency_type2.upper()
-    converted_amount = bot.currency.convert(currency_type1, currency_type2, float(amount))
-
-    await ctx.send('```%s %s → %0.2f %s```' % (amount, currency_type1, converted_amount, currency_type2))
 
 
 @bot.command(name='4chan', aliases=['4c', '4ch'])
